--- charIs.h
@@ -1,3 +1,0 @@
-int isDel(char ch);
-int isAlpha(char ch);
-int isDigit(char ch);
