+++ support.h
@@ -1,2 +1,5 @@
 #define LF 0xa
 void nullCheck(void* ptr);
+int isDel(char ch);
+int isAlpha(char ch);
+int isDigit(char ch);
